added model to different file and migration to database
